+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `fill-mask` pipeline
 - Added `image-classification` pipeline
 - Added `image-feature-extraction` pipeline
+- Added `object-detection` pipeline
 
 ## 1.0.3 (2024-08-29)